Implement full multi-user support with Username login and dynamic welcome message

+++ app/build/kspCaches/debug/backups/java/byRounds/1/ug/ac/ndejje/myapp/AIConversationDao_Impl.java
@@ -40,24 +40,25 @@ public final class AIConversationDao_Impl implements AIConversationDao {
       @Override
       @NonNull
       protected String createQuery() {
-        return "INSERT OR ABORT INTO `ai_conversations` (`id`,`role`,`content`,`timestamp`,`modeId`) VALUES (nullif(?, 0),?,?,?,?)";
+        return "INSERT OR ABORT INTO `ai_conversations` (`id`,`userId`,`role`,`content`,`timestamp`,`modeId`) VALUES (nullif(?, 0),?,?,?,?,?)";
       }
 
       @Override
       protected void bind(@NonNull final SupportSQLiteStatement statement,
           @NonNull final AIConversation entity) {
         statement.bindLong(1, entity.getId());
-        statement.bindString(2, entity.getRole());
-        statement.bindString(3, entity.getContent());
-        statement.bindLong(4, entity.getTimestamp());
-        statement.bindLong(5, entity.getModeId());
+        statement.bindLong(2, entity.getUserId());
+        statement.bindString(3, entity.getRole());
+        statement.bindString(4, entity.getContent());
+        statement.bindLong(5, entity.getTimestamp());
+        statement.bindLong(6, entity.getModeId());
       }
     };
     this.__preparedStmtOfDeleteAll = new SharedSQLiteStatement(__db) {
       @Override
       @NonNull
       public String createQuery() {
-        final String _query = "DELETE FROM ai_conversations";
+        final String _query = "DELETE FROM ai_conversations WHERE userId = ?";
         return _query;
       }
     };
@@ -82,12 +83,14 @@ public final class AIConversationDao_Impl implements AIConversationDao {
   }
 
   @Override
-  public Object deleteAll(final Continuation<? super Unit> $completion) {
+  public Object deleteAll(final int userId, final Continuation<? super Unit> $completion) {
     return CoroutinesRoom.execute(__db, true, new Callable<Unit>() {
       @Override
       @NonNull
       public Unit call() throws Exception {
         final SupportSQLiteStatement _stmt = __preparedStmtOfDeleteAll.acquire();
+        int _argIndex = 1;
+        _stmt.bindLong(_argIndex, userId);
         try {
           __db.beginTransaction();
           try {
@@ -105,9 +108,11 @@ public final class AIConversationDao_Impl implements AIConversationDao {
   }
 
   @Override
-  public Flow<List<AIConversation>> getConversation() {
-    final String _sql = "SELECT * FROM ai_conversations ORDER BY timestamp ASC";
-    final RoomSQLiteQuery _statement = RoomSQLiteQuery.acquire(_sql, 0);
+  public Flow<List<AIConversation>> getConversation(final int userId) {
+    final String _sql = "SELECT * FROM ai_conversations WHERE userId = ? ORDER BY timestamp ASC";
+    final RoomSQLiteQuery _statement = RoomSQLiteQuery.acquire(_sql, 1);
+    int _argIndex = 1;
+    _statement.bindLong(_argIndex, userId);
     return CoroutinesRoom.createFlow(__db, false, new String[] {"ai_conversations"}, new Callable<List<AIConversation>>() {
       @Override
       @NonNull
@@ -115,6 +120,7 @@ public final class AIConversationDao_Impl implements AIConversationDao {
         final Cursor _cursor = DBUtil.query(__db, _statement, false, null);
         try {
           final int _cursorIndexOfId = CursorUtil.getColumnIndexOrThrow(_cursor, "id");
+          final int _cursorIndexOfUserId = CursorUtil.getColumnIndexOrThrow(_cursor, "userId");
           final int _cursorIndexOfRole = CursorUtil.getColumnIndexOrThrow(_cursor, "role");
           final int _cursorIndexOfContent = CursorUtil.getColumnIndexOrThrow(_cursor, "content");
           final int _cursorIndexOfTimestamp = CursorUtil.getColumnIndexOrThrow(_cursor, "timestamp");
@@ -124,6 +130,8 @@ public final class AIConversationDao_Impl implements AIConversationDao {
             final AIConversation _item;
             final int _tmpId;
             _tmpId = _cursor.getInt(_cursorIndexOfId);
+            final int _tmpUserId;
+            _tmpUserId = _cursor.getInt(_cursorIndexOfUserId);
             final String _tmpRole;
             _tmpRole = _cursor.getString(_cursorIndexOfRole);
             final String _tmpContent;
@@ -132,7 +140,7 @@ public final class AIConversationDao_Impl implements AIConversationDao {
             _tmpTimestamp = _cursor.getLong(_cursorIndexOfTimestamp);
             final int _tmpModeId;
             _tmpModeId = _cursor.getInt(_cursorIndexOfModeId);
-            _item = new AIConversation(_tmpId,_tmpRole,_tmpContent,_tmpTimestamp,_tmpModeId);
+            _item = new AIConversation(_tmpId,_tmpUserId,_tmpRole,_tmpContent,_tmpTimestamp,_tmpModeId);
             _result.add(_item);
           }
           return _result;
